bitmap + metaphor - mutability = awesome?
So I started with the idea of splitting out the reachability data into some separate representation, and the concept of a continuity tester sounded good.  That separate representation turns out to be a bitmap, and from there, it turned out to be surprisingly straightforward to do a little fixed-point iteration to get to the end.